Services/eng-vault/app/requirements.txt
Flask>=2.0
gunicorn>=20.0
google-cloud-storage>=2.0
google-cloud-secret-manager>=2.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
